+++ js/util.js
@@ -1,0 +1,8 @@
+const getRandomPositiveInteger = (numberA, numberB) => {
+  const lower = Math.ceil(Math.min(Math.abs(numberA), Math.abs(numberB)));
+  const upper = Math.floor(Math.max(Math.abs(numberA), Math.abs(numberB)));
+  const result = Math.random() * (upper - lower + 1) + lower;
+  return Math.floor(result);
+};
+
+export {getRandomPositiveInteger};
